add snackbar for sucess and error based on operation

+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/addEditScreen/AddTaskScreen.kt
@@ -98,7 +98,7 @@ fun TudeeSnackBar(snackBarHostState: SnackbarHostState, isError: Boolean) {
                 .padding(horizontal = Theme.dimension.medium)
         ) { data ->
             SnackBar(
-                message = data.visuals.message,
+                message = data.visuals.message.toInt(),
                 snackBarStatus = if (isError) SnackBarStatus.ERROR else SnackBarStatus.SUCCESS
             )
         }
